+++ src/Presenters/AdminltePresenter.php
@@ -45,8 +45,8 @@ class AdminltePresenter extends BasePresenter
      */
     public function getMenuWithoutDropdownWrapper(MenuItem $item): string
     {
-        return '<li class="'.($item->isActive() ? 'active' : '').'">
-                    <a href="'.$item->getUrl().'" '.$item->getAttributes().'>
+        return '<li '.$item->getItemAttributes().'>
+                    <a href="'.$item->getUrl().'" '.$item->getLinkAttributes().'>
                         '.($item->icon ? '<i class="'.$item->icon.'"></i>' : '').'
                         <span>'.$item->title.'</span>
                     </a>