+++ src/Presenters/SidebarMenuPresenter.php
@@ -54,8 +54,8 @@ class SidebarMenuPresenter extends BasePresenter
      */
     public function getMenuWithoutDropdownWrapper(MenuItem $item): string
     {
-        return '<li class="'.($item->isActive() ? 'active' : '').'">
-                    <a href="'.$item->getUrl().'" '.$item->getAttributes().'>
+        return '<li '.$item->getItemAttributes().'>
+                    <a href="'.$item->getUrl().'" '.$item->getLinkAttributes().'>
                         '.($item->icon ? '<i class="'.$item->icon.'"></i>' : '').'
                         '.$item->title.'
                     </a>